# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: BetPulse Sports Betting Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /wallet/deposit:
    post:
      operationId: deposit
      tags: [wallet]
      summary: Deposit funds
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepositBody"
      responses:
        "200":
          description: Deposit successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WalletBalance"
  /wallet/transactions:
    get:
      operationId: getTransactions
      tags: [wallet]
      summary: Get transaction history
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
...
components:
  schemas:
...
    WalletBalance:
      type: object
      required: [balance, totalDeposited, totalWagered, totalWon]
      properties:
        balance:
          type: number
        totalDeposited:
          type: number
        totalWagered:
          type: number
        totalWon:
          type: number
    DepositBody:
      type: object
      required: [amount]
      properties:
        amount:
          type: number
    Transaction:
      type: object
      required: [id, type, amount, balanceAfter, description, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [deposit, withdrawal, bet_placed, bet_won, bet_lost, bonus]
        amount:
          type: number
        balanceAfter:
          type: number
        description:
          type: string
        createdAt:
          type: string
...